+++ Musique/app/src/main/java/com/randsoft/apps/musique/renderdata/Text.kt
@@ -4,6 +4,6 @@ data class Text(
     var text: String = "",
     var x: Float = 0.0f,
     var y: Float = 0.0f,
-    var paint: Paint = Paint(0x00000000.toInt())
+    var paint: Paint = Paint(0x00000000)
 ) {
 }
